fitprint_server/requirements.txt
fastapi[standard]==0.120.0
boto3==1.40.59
python-dotenv==1.1.1
